nlp/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
python-dotenv==1.0.0
requests==2.31.0
pytest==8.4.2
openai>=1.0.0
cerebras-cloud-sdk>=1.0.0
